Revert "[software] enable floating downscaling"
This reverts commit ad1180e4c3f9f1a881a946ed93c7a24dfcfd4bc7.

+++ src/aliceVision/image/resampling.hpp
@@ -47,7 +47,7 @@ namespace image {
   }
 
   template <typename SamplerType, typename Image>
-  void downscaleImage(const Image& src, Image& out, float downscale)
+  void downscaleImage(const Image& src, Image& out, int downscale)
   {
       const int new_width = src.Width() / downscale;
       const int new_height = src.Height() / downscale;
@@ -67,7 +67,7 @@ namespace image {
   }
 
   template <typename ImageType>
-  void downscaleImageInplace(ImageType& inout, float downscale)
+  void downscaleImageInplace(ImageType& inout, int downscale)
   {
       if(downscale <= 1)
           return;
